refactor(form-documents): rename ingestion service to form-documents
Reflects that this service owns both extraction (PDF→data) and filling
(data→PDF) — intent over mechanism per P1.

- Rename src/services/ingestion/ → src/services/form-documents/
- Rename pdf-extractor.ts → extraction.ts within the module
- Delete stale top-level pdf-extractor.ts and extraction-schemas.ts
- Update all imports across src/ and test/
- Install pdf-lib dependency for upcoming PDF filling capability
