+++ package.json
@@ -28,6 +28,7 @@
     "hono": "^4.0.0",
     "markdown-it": "^14.1.1",
     "markdown-it-task-lists": "^2.1.1",
+    "pdf-lib": "^1.17.1",
     "zod": "^4.3.6"
   },
   "devDependencies": {
