+++ test/project-service.test.ts
@@ -11,7 +11,7 @@ import {
 } from '../src/services/errors'
 import type { FormProjectRepo } from '../src/services/form-project-repo'
 import { createFormProjectRepo } from '../src/services/form-project-repo'
-import type { PdfExtractor } from '../src/services/pdf-extractor'
+import type { PdfExtractor } from '../src/services/form-documents/extraction'
 import {
   createProjectService,
   type ProjectService,
